Refactor SimilarIdeas imports to use SimilarInputs context and components
