+++ front/app/containers/IdeasNewPage/SimilarInputs/InputDetailView.tsx
