+++ front/app/containers/IdeasNewPage/SimilarInputs/messages.ts
